# Optimize short string deserialization