Significantly improve reading short strings. This greatly decreases the time to read frame_id from header messages.

_read last field of std_msgs/Header_ benchmark

on main:

```
Running "std_msgs/Header - read last field" suite...
Progress: 50%

  Reg:
    1 152 386 ops/s, ±1.14% 
Progress: 100%

  Reg:
    1 152 386 ops/s, ±1.14%   | fastest

  Lazy:
    319 905 ops/s, ±2.42%     | slowest, 72.24% slower
```

on pr

```
Running "std_msgs/Header - read last field" suite...
Progress: 50%

  Reg:
    1 068 996 ops/s, ±1.15% 
Progress: 100%

  Reg:
    1 068 996 ops/s, ±1.15%   | slowest, 76.82% slower

  Lazy:
    4 611 377 ops/s, ±0.80%   | fastest

```